+++ start/start/src/main/java/com/house/start/controller/ConsumerController.java
@@ -33,8 +33,7 @@ public class ConsumerController {
     private final ItemService itemService;
     private final OrderService orderService;
 
-    private Logger logger = LoggerFactory.getLogger(ConsumerController.class);
-    private final ServletContext servletContext;
+
 
     // 물건 리스트
     @GetMapping("/list")
@@ -76,137 +75,12 @@ public class ConsumerController {
         return "redirect:/list";
     }
 
-
-    // 장바구니
-//    @GetMapping("/cart")
-//    public String getCarts(Model model){
-//        List<Item> itemList = consumerService.findByCart(ItemStatus.CART);
-//        model.addAttribute("itemList", itemList);
-//        return "cart";
-//    }
-
-
     // 배송 완료
     @PutMapping("/seller/deliveries/{id}/completed")
     public void getAllDelivery(@PathVariable Long id){
         consumerService.completeDelivery(id);
     }
 
-    // 커뮤니티 목록
-    @GetMapping("/community")
-    public String getAllPost(Model model) {
-        String realPath = servletContext.getRealPath("/resources");
-        logger.info("realPath:  "+realPath);
-
-        List<Post> posts = consumerService.getAllPost();
-        model.addAttribute("postList", posts);
-        return "consumer/post_list";
-    }
-
-    // 글 조회
-    @GetMapping("/community/{id}")
-    public String getOnePost(@PathVariable Long id,
-                             @SessionAttribute(name = SessionConstants.LOGIN_MEMBER, required = false) Consumer loginConsumer,
-                             Model model, HttpServletRequest  request) {
-        Post post = consumerService.getOnePost(id);
-        List<Like> likes = post.getLikes();
-        // 글 조회수 update
-        consumerService.updateView(id);
-
-        if(loginConsumer!=null) {
-            Boolean flag = false;
-            for(Like like : likes) {
-                if(like.getConsumer().getId()==loginConsumer.getId()) {
-                    flag = true;
-                    break;
-                }
-            }
-            model.addAttribute("liked", flag);
-        }
-
-        model.addAttribute("post", post);
-        model.addAttribute("likes", likes);
-        model.addAttribute("comments", post.getComments());
-
-        return "consumer/consumer_postDetail";
-    }
-
-
-    // 글 작성 페이지
-    @GetMapping("/community/new")
-    public String getNewPost(@SessionAttribute(name = SessionConstants.LOGIN_MEMBER, required = false) Consumer loginConsumer,
-                             Model model) {
-        if(loginConsumer == null) {
-            return "redirect:/login";
-        }
-        model.addAttribute("post", new PostForm());
-        return "consumer/consumer_newPost";
-    }
-
-    // 글 작성
-    @PostMapping("/community/write")
-    public String postUser(@ModelAttribute PostForm post,
-                           @SessionAttribute(name = SessionConstants.LOGIN_MEMBER, required = false) Consumer loginConsumer,
-                           HttpServletRequest request) throws IOException {
-
-        logger.info(post.getContents());
-        logger.info(String.valueOf(post.getPhoto()));
-
-        UploadFile uploadFile = fileStore.storeFile(post.getPhoto(), request);
-
-        Post newPost = Post.builder()
-                .contents(post.getContents())
-                .uploadFile(uploadFile)
-                .consumer(loginConsumer)
-                .postDate(LocalDateTime.now())
-                .build();
-
-        consumerService.savePost(newPost);
-
-        return "redirect:/community";
-    }
-    /**
-     * 글 -> 좋아요 누르기
-     * **/
-    @PostMapping("/community/{id}/likes")
-    public String putLikes(@PathVariable String id, @SessionAttribute(name = SessionConstants.LOGIN_MEMBER, required = false) Consumer loginConsumer) {
-        if(loginConsumer == null) {
-            return "redirect:/login";
-        }
-
-        Long likeId = consumerService.putLikes(Long.valueOf(id), loginConsumer);
-        System.out.println("Like Create No: "+likeId);
-
-
-        return "redirect:/community/"+id;
-    }
-
-
-    /**
-     * 댓글
-     * **/
-    // 댓글 작성
-    @PostMapping("/community/{id}/comments/write")
-    public String postComment(@PathVariable String id, @RequestParam String contents, @SessionAttribute(name = SessionConstants.LOGIN_MEMBER, required = false) Consumer loginConsumer) {
-        consumerService.saveComment(id, contents, loginConsumer);
-        return "redirect:/community/"+id;
-    }
-
-    // 댓글 수정
-    @PutMapping("/comments/{id}/put")
-    public String putComment(@PathVariable String id, @RequestBody CommentForm commentForm, Model model) {
-        consumerService.updateComment(commentForm.getId(), commentForm.getContent());
-        model.addAttribute("ACCESS", "SUCCESS");
-        return "redirect:/community/"+commentForm.getPostId();
-    }
-
-    // 댓글 삭제
-    @DeleteMapping("/comments/{id}/delete")
-    public String deleteComment(@PathVariable Long id, Model model) {
-        consumerService.deleteComment(id);
-        model.addAttribute("ACESS", "SUCCESS");
-        return "redirect:/community/";
-    }
 
     /**
      *  상품 목록 페이지
