+++ start/start/src/main/java/com/house/start/domain/dto/Cart/CartItemDto.java
@@ -1,0 +1,11 @@
+package com.house.start.domain.dto.Cart;
+
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class CartItemDto {
+
+}
